refactor: module input
